explicitly Long.toDouble

+++ json/src/main/scala/smile/json/JsValue.scala
@@ -222,7 +222,7 @@ case class JsLong(value: Long) extends JsValue with Ordered[JsLong] {
   override def asBoolean: Boolean = value != 0
   override def asInt: Int = value.toInt
   override def asLong: Long = value
-  override def asDouble: Double = value
+  override def asDouble: Double = value.toDouble
 
   override def compare(that: JsLong): Int = {
     value.compare(that.value)
@@ -254,7 +254,7 @@ case class JsCounter(value: Long) extends JsValue with Ordered[JsCounter] {
   override def asBoolean: Boolean = value != 0
   override def asInt: Int = value.toInt
   override def asLong: Long = value
-  override def asDouble: Double = value
+  override def asDouble: Double = value.toDouble
 
   override def compare(that: JsCounter): Int = {
     value.compare(that.value)
@@ -364,7 +364,7 @@ case class JsDate(value: LocalDate) extends JsValue with Ordered[JsDate] {
     * The Epoch Day count is a simple incrementing count of days
     * where day 0 is 1970-01-01 (ISO).
     */
-  override def asDouble: Double = value.toEpochDay
+  override def asDouble: Double = value.toEpochDay.toDouble
 
   override def compare(that: JsDate): Int = {
     value.compareTo(that.value)
@@ -408,7 +408,7 @@ case class JsTime(value: LocalTime) extends JsValue with Ordered[JsTime] {
   /** Converts this time as nanos of day, from 0 to 24 * 60 * 60 * 1,000,000,000 - 1. */
   override def asLong: Long = value.toNanoOfDay
   /** Converts this time as nanos of day, from 0 to 24 * 60 * 60 * 1,000,000,000 - 1. */
-  override def asDouble: Double = value.toNanoOfDay
+  override def asDouble: Double = value.toNanoOfDay.toDouble
 
   override def compare(that: JsTime): Int = {
     value.compareTo(that.value)
@@ -451,7 +451,7 @@ case class JsDateTime(value: LocalDateTime) extends JsValue with Ordered[JsDateT
   /** Converts this date-time to the number of seconds from the epoch of 1970-01-01T00:00:00Z. */
   override def asLong: Long = value.toEpochSecond(ZoneOffset.UTC)
   /** Converts this date-time to the number of seconds from the epoch of 1970-01-01T00:00:00Z. */
-  override def asDouble: Double = value.toEpochSecond(ZoneOffset.UTC)
+  override def asDouble: Double = value.toEpochSecond(ZoneOffset.UTC).toDouble
 
   override def compare(that: JsDateTime): Int = {
     value.compareTo(that.value)
@@ -485,7 +485,7 @@ case class JsTimestamp(value: Timestamp) extends JsValue with Ordered[JsTimestam
   }
   override def asDateTime: LocalDateTime = value.toLocalDateTime
   override def asLong: Long = value.getTime
-  override def asDouble: Double = value.getTime
+  override def asDouble: Double = value.getTime.toDouble
 
   override def compare(that: JsTimestamp): Int = {
     value.compareTo(that.value)
